+++ esmerald/conf/global_settings.py
@@ -875,7 +875,9 @@ class EsmeraldAPISettings(BaseSettings):
         return None
 
     @property
-    def static_files_config(self) -> Optional[StaticFilesConfig]:
+    def static_files_config(
+        self,
+    ) -> Union[StaticFilesConfig, list[StaticFilesConfig], tuple[StaticFilesConfig, ...], None]:
         """
         An instance of [StaticFilesConfig](https://esmerald.dev/configurations/staticfiles/).